Prevents MySQL driver to failing when version table alredy exists
The MySQL driver is using `CREATE TABLE IF NOT EXISTS`, which causes
the MySQL database to raise warnings. The golang driver collects
the warnings and return a composite object (mysql.MySQLWarnings) as
an error that needs to be properly handled.

This change stops the driver from failing in case there are only
warnings.